refactor(json): rename jq module to format
